// app/src/pages/project/__generated__/TraceCountTimeSeriesQuery.graphql.ts
export type TimeRange = {
  end?: string | null;
  start?: string | null;
};
export type TimeBinConfig = {
  scale?: TimeBinScale;
  utcOffsetMinutes?: number;
};
export type TraceCountTimeSeriesQuery$variables = {
  projectId: string;
  timeBinConfig: TimeBinConfig;
  timeRange: TimeRange;
};
export type TraceCountTimeSeriesQuery$data = {
  readonly project: {
    readonly traceCountByStatusTimeSeries?: {
      readonly data: ReadonlyArray<{
        readonly errorCount: number;
        readonly okCount: number;
        readonly timestamp: string;
      }>;
    };
  };
};
export type TraceCountTimeSeriesQuery = {
  response: TraceCountTimeSeriesQuery$data;
  variables: TraceCountTimeSeriesQuery$variables;
};